Final cleanup and fixes
