+++ constants.json
@@ -5,5 +5,6 @@
   "googleAnalyticsId": "UA-168767157-1",
   "emoteSize": 24,
   "chatRoomId": "!doYmXmbbmJCuwjKvff:woke.net",
-  "announcementsRoomId": "!zxoqtUJzGHtPOeXuaH:woke.net"
+  "announcementsRoomId": "!zxoqtUJzGHtPOeXuaH:woke.net",
+  "maximumMessageSize": 500
 }
